compute test -> if config is correct

+++ include/classifier/classifier.h
@@ -5,6 +5,8 @@
 #ifndef NBC4GPU_CLASSIFIER_H
 #define NBC4GPU_CLASSIFIER_H
 
+#include <vector>
+
 namespace nbc4gpu {
   /**
    * Dummy class for project test
@@ -13,7 +15,7 @@ namespace nbc4gpu {
     using LearnClassifier = int;  // TODO
     using Predictor = int;        // TODO
 
-    using Attribute = float;
+    using Attribute = double;
     using Record = std::vector<Attribute>; //!< row containing value in each column
 
     using Column = std::vector<Attribute>; //!< contains value for each row
@@ -31,6 +33,7 @@ namespace nbc4gpu {
      */
     double predict(const Record &record);
   private:
+
     int field;
   };
 } // namespace nbc4gpu
